move optional image decoding into ImageWithNounFallback + add custom handling for different URI sources
