+++ compositions/NFTPage/NFTPageHero.tsx
@@ -2,11 +2,9 @@ import { ImageWithNounFallback } from 'components'
 
 import { useToken } from 'hooks/useToken'
 
-import { TypeSafeDao } from 'validators/dao'
 import { TypeSafeToken } from 'validators/token'
 
 import { cardImageWrapper } from '@media/NftMedia.css'
-import { useOptionalImageURIDecode } from '@media/hooks/useImageURIDecode'
 import { Box, BoxProps } from '@zoralabs/zord'
 
 import { nftPageHero } from './NFTPage.css'
@@ -33,8 +31,6 @@ export function NFTPageHeroComponent({
   token,
   ...props
 }: Omit<NFTPageHeroProps, 'collectionAddress' | 'tokenId'> & { token: TypeSafeToken }) {
-  const srcImg = useOptionalImageURIDecode(token) // Handle non-base64 SVGs by decoding URI. This should be replaced when handled properly API-side
-
   return (
     <Box
       w="100%"
@@ -43,11 +39,7 @@ export function NFTPageHeroComponent({
       overflow="hidden"
       {...props}
     >
-      <ImageWithNounFallback
-        tokenContract={token.collectionAddress}
-        tokenId={token.tokenId}
-        srcImg={srcImg}
-      />
+      <ImageWithNounFallback token={token} />
     </Box>
   )
 }
